test, docs: Functional, test code and comments
This version of the translator is a working one, it has been proven with
the codes 1 to 3, from the repository of the class.
Also comments for the code were added in all of the algorithms,
variables and cycles to explain the functionality of each code segment.
At the bottom, a test code was added, simulating a list being passed to
the function, and the returned list by the function is printed by a for
iterator.

-beq and bne BranAddr two's complement issue corrected.