fix, style: Fixed sb and lb instructions
There was an issue with the instructions lb and sb, it didn't converted
the values of imm correctly, so, the respective lines in translator.py
were edited due to fix this problem.

Also, all the codes were formatted using the pylint format helper. We
wanted the project to be called ISA.py.

Added an exit condition to the program, this avoiding the program to
continue executing when an error in the code is detected, and to prevent
the program to display error messages to the user.